Initial revision of a CPP implementation of a DLX assembler.
At the moment it supports parsing a single line and extracting the label,
instruction, registers and comment. It lacks is lacking the ability to
recongise what to expect after the instruction (i.e which instruction format).
Error handling is non-existent, so giving it an invalid string to parse
won't result in any kind of diagnostics.